externalized the SpineDebugRenderer
